Add files via upload
My Final Year Project; a initiative to help farmers to rent their farming equipment when not required and add an extra source to their income. They can operate the application through their cell phone and can even manage the payment through that application. Their are various functionalities listed as well for the farmers through which they can operate the application.